backend/requirements.txt
fastapi==0.109.1
python-multipart==0.0.9
uvicorn==0.27.0
pycryptodome==3.20.0
pyuploadcare==5.0.0
python-dotenv==1.0.1
requests==2.31.0